+++ src/components/CartList/CartList.style.ts
@@ -1,0 +1,26 @@
+import styled from "styled-components";
+
+export const CartListTitle = styled.div`
+  font-style: normal;
+  font-weight: 400;
+  font-size: 20px;
+  line-height: 33px;
+
+  letter-spacing: 0.5px;
+
+  margin-top: 40px;
+`;
+
+export const CartListController = styled.div`
+  display: flex;
+`;
+
+export const CartListCheckCounter = styled.div`
+  font-style: normal;
+  font-weight: 400;
+  font-size: 16px;
+  line-height: 20px;
+
+  letter-spacing: 0.5px;
+  margin: 0px 15px 0px 15px;
+`;
