+++ src/components/CartList/CartList.tsx
@@ -1,0 +1,26 @@
+import { useRecoilValue } from "recoil";
+import useCart from "../../hooks/useCart";
+import { cartCountSelector } from "../../recoil/cartAtoms";
+import CartItem from "../CartItem";
+import { CartListCheckCounter, CartListController, CartListTitle } from "./CartList.style";
+
+function CartList() {
+  const { cartList } = useCart();
+  const cartCount = useRecoilValue(cartCountSelector);
+
+  return (
+    <div>
+      <CartListTitle>든든배송 상품 ({cartCount}개)</CartListTitle>
+      {cartList.map((cart) => (
+        <CartItem key={cart.id} cart={cart} />
+      ))}
+      <CartListController>
+        <input type='checkbox' />
+        <CartListCheckCounter>전체선택 (2/{cartCount})</CartListCheckCounter>
+        <div><button>선택삭제</button></div>
+      </CartListController>
+    </div>
+  );
+}
+
+export default CartList;
