+++ src/components/CartList/index.tsx
@@ -1,0 +1,3 @@
+import CartList from './CartList';
+
+export default CartList;
